src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.12.0=pypi_0
appnope=0.1.2=py37hecd8cb5_1001
argon2-cffi=20.1.0=py37h9ed2024_1
astunparse=1.6.3=pypi_0
async_generator=1.10=py37h28b3542_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
bzip2=1.0.8=hc929b4f_4
ca-certificates=2021.5.25=hecd8cb5_1
cachetools=4.2.1=pypi_0
cairo=1.16.0=h0ab9d94_1001
certifi=2021.5.30=py37hecd8cb5_0
cffi=1.14.5=py37h2125817_0
chardet=4.0.0=pypi_0
cloudpickle=1.6.0=py_0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37haf1e3a3_0
dask-core=2021.4.0=pyhd3eb1b0_0
dbus=1.13.18=h18a8e69_0
decorator=5.0.6=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py37_0
expat=2.3.0=h23ab428_2
ffmpeg=4.0=h01ea3c9_0
flatbuffers=1.12=pypi_0
fontconfig=2.13.1=hd23ceaa_1004
freetype=2.10.4=ha233b18_0
fsspec=0.9.0=pyhd3eb1b0_0
gast=0.3.3=pypi_0
gettext=0.21.0=h7535e17_0
glib=2.68.0=hdf23fa2_0
google-auth=1.30.0=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.13=h12caacf_1001
grpcio=1.32.0=pypi_0
h5py=2.10.0=pypi_0
harfbuzz=1.8.8=hb8d4a28_0
hdf5=1.10.2=hc401514_3
icu=58.2=h0a44026_3
idna=2.10=pypi_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=3.10.0=py37hecd8cb5_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2019.4=233
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.22.0=py37h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jasper=2.0.14=h636a363_1
jedi=0.17.0=py37_0
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=he5867d9_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py37_7
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.7.1=py37hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py37h23ab428_0
lcms2=2.12=hf1fd2bf_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libopencv=3.4.2=h7c891bd_1
libopus=1.3.1=hc929b4f_1
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
libvpx=1.7.0=h378b8a2_0
libxml2=2.9.10=h7cdb67c_3
llvm-openmp=10.0.0=h28b9765_0
locket=0.2.1=py37hecd8cb5_1
lz4-c=1.9.3=h23ab428_0
markdown=3.3.4=pypi_0
markupsafe=1.1.1=py37h1de35cc_0
matplotlib=3.3.4=py37hecd8cb5_0
matplotlib-base=3.3.4=py37h8b3ea08_0
mistune=0.8.4=py37h1de35cc_0
mkl=2019.4=233
mkl-service=2.3.0=py37h9ed2024_0
mkl_fft=1.3.0=py37ha059aab_0
mkl_random=1.1.1=py37h959d312_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py37_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.2=py37_1
ninja=1.10.2=hf7b0b51_1
notebook=6.3.0=py37hecd8cb5_0
numpy=1.19.2=py37h456fd55_0
numpy-base=1.19.2=py37hcfb5961_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1k=h0d85af4_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.3=py37hb2f4e1b_0
pandoc=2.12=hecd8cb5_0
pandocfilters=1.4.3=py37hecd8cb5_1
parso=0.8.2=pyhd3eb1b0_0
partd=1.1.0=py_0
pcre=8.44=hb1e8313_0
pexpect=4.8.0=pyhd3eb1b0_3
pickle5=0.0.11=py37h9bfed18_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py37h5270095_0
pip=21.1=pypi_0
pixman=0.38.0=h01d97ff_1003
prometheus_client=0.10.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
prompt_toolkit=3.0.17=hd3eb1b0_0
protobuf=3.15.8=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-opencv=3.4.2=py37h7c891bd_1
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pygments=2.8.1=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h655552a_2
pyrsistent=0.17.3=py37haf1e3a3_0
python=3.7.10=h88f2d9e_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.7=1_cp37m
pytorch=1.7.1=py3.7_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py37haf1e3a3_2
pyyaml=5.4.1=py37h9ed2024_1
pyzmq=20.0.0=py37h23ab428_1
qt=5.9.7=h468cd18_1
qtconsole=5.0.3=pyhd3eb1b0_0
qtpy=1.9.0=py_0
readline=8.1=h9ed2024_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-image=0.18.1=py37hb2f4e1b_0
scikit-learn=0.24.1=py37hb2f4e1b_0
scipy=1.6.2=py37h2515648_0
seaborn=0.11.1=pyhd3eb1b0_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py37hecd8cb5_0
sip=4.19.8=py37h0a44026_0
six=1.15.0=py37hecd8cb5_0
sqlite=3.35.4=hce871da_0
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.0=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.4.1=pypi_0
tensorflow-estimator=2.4.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.4=py37hecd8cb5_0
testpath=0.4.4=pyhd3eb1b0_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.10.1=py37h0cf3a3e_2
tk=8.6.10=hb0a8c7a_0
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cpu
tornado=6.1=py37h9ed2024_0
tqdm=4.59.0=pyhd3eb1b0_1
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py37_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zeromq=4.3.4=h23ab428_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
